div file to files and edit from tkinter import * to import tkinter as tk
